requirements.txt
Django
djangorestframework
django-cors-headers
pandas
openpyxl
psycopg2
psycopg